SD E2E-style tests
