+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
 	"bb-imager-cli", "bb-flasher-pb2-mspm0",
 	"bb-imager-gui", "bb-imager-service",
 	"xtask", "bb-flasher-bcf", "bb-flasher-sd", "iced-loading", "bb-downloader", "bb-config", "bb-helper", "bb-flasher-dfu",
+	# removed non-standard e2e-tests crate; E2E-style tests will live under tests/e2e
 ]
 
 [profile.release]
