--- e2e-tests/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "e2e-tests"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-repository.workspace = true
-license.workspace = true
-publish = false
-
-[dependencies]
-tokio = { version = "1", features = ["full"] }
-uuid = { version = "1", features = ["v4"] }
-xz2 = "0.1"
-bb-flasher = { path = "../bb-flasher", default-features = false }
-bb-helper = { path = "../bb-helper" }
-tokio-util = "0.7"
-
-[features]
-default = []
-sd = ["bb-flasher/sd"]
-bcf = ["bb-flasher/bcf"]
-bcf_msp430 = ["bb-flasher/bcf_msp430"]
-dfu = ["bb-flasher/dfu"]
-all = ["sd", "bcf", "bcf_msp430", "dfu"]
-
